move input and output handling to the same module
